emacs: fix siteVersionDir path
siteVersionDir isn't a full path, it's just the name of the version
directory entry in $out/share/emacs, e.g. "27.1", so since
d1b0eef9b51 ("emacs: Don't use interpolation for version"), we've been
trying to remove a non-existent directory.  This would have been
caught if we hadn't been giving -f to rm unnecessarily, because -f
suppresses errors.  As well as fixing the path, I've removed the -f
from rm.  Doing this to the line above as well revealed that rm-ing
$out/var was no longer necessary, since nothing has been put there
since ac23a7c459d8 ("emacs: 25.3 → 26.1") -- it would now only be
created if we set the --with-gameuser configure option, which we
don't.