zcash: 2.1.1-1 -> 4.1.1; fix build
This is pretty much a repackaging, because a lot has changed about
zcash in two major versions.  Lots of libraries are no longer
necessary, and librustzcash is now part of the zcash package as
opposed to being a library supposed to be built seperately.

The whole thing is a buildRustPackage because that's easier than
trying to emulate buildRustPackage's dependency setup behaviour inside
a normal mkDerivation.  So that the normal zcash build process is
followed, the actual configuring and building uses the normal stdenv
behaviour.