platformio: Enable shell completion generation
The `platformio system completion install` command relies on the
click-completion python package to derive completions from the current
click argument logic. When this command is ran, it currently errors on
pip exceptions. If a nix-shell is used to run this command with
click-completion installed, shell completion is broken when outside
of the nix-shell.