cc-wrapper: disable dynamic linker for darwin
Disable the dynamic linker if the platform is darwin. On darwin the
dynamic linker cannot be passed using the -dynamic-linker argument, but
needs to be explicitly set via environment variables. Passing the linker
via -dynamic-linker results in an error:

  ld: unknown option: -dynamic-linker=/usr/lib/dyld

Pass a isdarwin variable to the build script and disable the linker on
darwin systems.