waffle: Add support for non-linux platforms
Waffle works perfectly well on macOS and Windows. Because the meson
options do proper dependency checking already, we can just leave them
out. This has the additional benefit of allowing gbm, surfaceless_egl
and x11_egl supoort to be built on Linux, which are useful in
conjunction with the piglit test suite.