nixos/libinput: make every option optional
The libinput module nowadays set some options as optional (i.e.: accepts
null) and other options as required (i.e.: if not set it will be set
with a default).

However, this module, since PR #73785, applies to all input devices. And
this is an issue since many of the options here only make sense to
touchpad (for example, NaturalScrolling). Worse than that, once any option is
in this configuration it cannot be overriden by any of the existing mechanisms
to override, like:

- `services.xserver.extraConfig`
- `services.xserver.inputClassSections`

The only one that works is `services.xserver.libinput.additionalOptions`, but
this one does not support applying options to many targets. You can set
`additionalOptions` with `MatchIsTouchpad "on"`, for example, but now
you cannot apply options in other target unless you use one of the
options above.

So what this PR does is convert every option to optional, and also
setting to `null` any option that is already default in libinput itself.
This way, most options will be available to override using
`services.xserver.extraConfig` or `services.xserver.inputClassSections`.

Closes https://github.com/NixOS/nixpkgs/issues/75007.