pythonPackages.demjson: fix tests, enable on Python 3.x
In commit 6ba044c1667a4966a69debe73940972307225301, the demjson package was
disabled on Python 3.x with the comment that it doesn't seem to support any
Python 3.x versions.  But looking at the upstream repository, they do seem to
attempt to support Python 3 -- it turns out the failure on our end was caused by
some issue with trying to run `setup.py test` on a 2to3-using codebase with no
test suite (?).

In any case, this package's test suite doesn't seem to use the setuptools
mechanism, so in this commit I override the checkPhase to run the upstream tests
in the correct way.  This fixes the build on all Python versions.

EDIT 2021-01-08: rebased on top of PR #108378 which had explicitly disabled the
tests on all Python versions.