brave: add update script
The Brave package often seems to get very outdated. This is bad for a
browser, where vulnerabilities are high impact.

This change adds an update script, so that r-ryantm will suggest
updates. We find the latest version using their Debian package
database (since we are using the Debian package anyway).