samba4: Properly use Python 3 and fix PYTHONPATH
Some tools (especially AD tools) don't work without dnspython and tdb.
This change adds them to $PYTHONPATH.

Also this PR fixes the ugly Python 3 override.

To test this, it's not enough to just build the package, but you should
rather try a tool (like samba_dnsupdate) which is built on Python. If it
does not return an ImportError, this change worked.