nixos/tests/dovecot: enable dovecot_pigeonhole plugin
This plugin is used commonly enough that we should ensure it still
builds (and dovecot works) after loading it.

This is not yet perfect as we aren't testing any of it's functionality
but at least we ensure that dovecot continues to do the regular job.

(cherry picked from commit 653f805f098de519f0204e1eed00b26e80d03ce2)