nixos/redshift: use config file instead of passing parameters
Not every option is exposed by redshift/gammastep parameters, for
example gamma options are only exposed in configuration file. So this PR
refactors this module to generate a configuration file and pass it to
the redshift/gammastep using -c parameter.

Also, it implements fade/gamma options not available in the previous
version of this module (since it was not exposed by parameters).

This is a slightly breaking change, since extraOptions was renamed to
extraConfig and now accept a attribute set instead of a list of strings.