far2l: 2019-12-14 -> 2020-12-30
This change is required to enable running far2l on GUI-less linux, older version were bound to GTK+ and failed to start.
The colorer has been updated since in the far2l repo to a fresh version, so probably patching the schemes is not needed.
Added missing packages which are now needed to compile successfully.