nixos/redshift: use config file instead of passing parameters
Not every option is exposed by redshift/gammastep parameters, for
example gamma options are only exposed in configuration file. So this PR
refactors this module to generate a configuration file and pass it to
the redshift/gammastep using -c parameter.

This is a breaking change since there is no support for some of the
older options like `extraConfig`. Also, the way to pass settings
completely changed (but there is some aliases).