doc: add "prefer lib over stdenv.lib" convention
I think we should have something in the manual people can point to
about this, to avoid rehashing it over and over in PRs.  "stdenv.lib"
makes it look like lib is part of stdenv, which it isn't, and makes it
even more confusing as a newcomer to figure out what stdenv is (and
isn't).