freeimage: 3.18.0 -> unstable-2020-07-04, unbundle dependencies, fix aarch64 support
The unreleased version of freeimage contains many important fixes,
amongst others CVEs, and is taken from the svn repository (r1859).

We also unbundle all the dependencies to make it fit into the Nixpkgs ecosystem.

All the changes needed to unbundle and make it compile with Nix is contained in unbundle.diff.

Fixes #77653
Replaces #77655