Fix header for generated Dhall documentation
By default, `dhall-docs` uses the name of the input directory
as the initial component of the documentation header.  However,
since the input directory is built using Nix the header contains
the Nix store hash in the name, which then appears in the
generated documentation.

The fix is to override this default behavior by supplying the
`--package-name` flag to `dhall-docs`.