libvirt: correct localstatedir
It was supposed to be '/var' rather than '/var/lib'. This fixes an issue
where some tools don't connect to the right socket (because they were
connecting to '$localstatedir/run/libvirt...' instead of
'$runstatedir/libvirt...').

Also change runstatedir to just be '/run' since it avoids a log line
complaining about that.